requirements.txt
graphiti_core>=0.24.3
python-dotenv>=1.0.0
neo4j>=5.15.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.6
openai>=1.0.0
httpx>=0.25.0
